(PC-33503)[PRO] feat: refacto and remove one step
