+++ pro/src/pages/CollectiveOffer/CollectiveOfferLayout/CollectiveEditionOfferNavigation/CollectiveEditionOfferNavigation.module.scss
