+++ pro/src/pages/CollectiveOffer/CollectiveOfferLayout/CollectiveEditionOfferNavigation/CollectiveEditionOfferNavigation.tsx
@@ -1,32 +1,29 @@
 import cn from 'classnames'
 import { useState } from 'react'
 import { useSelector } from 'react-redux'
-import { useLocation, useNavigate } from 'react-router-dom'
-import { useSWRConfig } from 'swr'
+import { useNavigate, useLocation } from 'react-router-dom'
+import { mutate } from 'swr'
 
 import { api } from 'apiClient/api'
 import {
-  CollectiveOfferAllowedAction,
-  CollectiveOfferDisplayedStatus,
   CollectiveOfferStatus,
+  CollectiveOfferDisplayedStatus,
   CollectiveOfferTemplateAllowedAction,
+  CollectiveOfferAllowedAction,
   GetCollectiveOfferResponseModel,
   GetCollectiveOfferTemplateResponseModel,
 } from 'apiClient/v1'
 import { useAnalytics } from 'app/App/analytics/firebase'
 import {
-  GET_COLLECTIVE_OFFER_QUERY_KEY,
   GET_COLLECTIVE_OFFER_TEMPLATE_QUERY_KEY,
+  GET_COLLECTIVE_OFFER_QUERY_KEY,
 } from 'commons/config/swrQueryKeys'
 import {
   Events,
   COLLECTIVE_OFFER_DUPLICATION_ENTRIES,
 } from 'commons/core/FirebaseEvents/constants'
 import { NOTIFICATION_LONG_SHOW_DURATION } from 'commons/core/Notification/constants'
-import {
-  isCollectiveOffer,
-  isCollectiveOfferTemplate,
-} from 'commons/core/OfferEducational/types'
+import { isCollectiveOffer } from 'commons/core/OfferEducational/types'
 import { computeURLCollectiveOfferId } from 'commons/core/OfferEducational/utils/computeURLCollectiveOfferId'
 import { createOfferFromTemplate } from 'commons/core/OfferEducational/utils/createOfferFromTemplate'
 import { duplicateBookableOffer } from 'commons/core/OfferEducational/utils/duplicateBookableOffer'
@@ -36,7 +33,6 @@ import { selectCurrentOffererId } from 'commons/store/offerer/selectors'
 import { isActionAllowedOnCollectiveOffer } from 'commons/utils/isActionAllowedOnCollectiveOffer'
 import { ArchiveConfirmationModal } from 'components/ArchiveConfirmationModal/ArchiveConfirmationModal'
 import { canArchiveCollectiveOfferFromSummary } from 'components/ArchiveConfirmationModal/utils/canArchiveCollectiveOffer'
-import { Step, Stepper } from 'components/Stepper/Stepper'
 import fullArchiveIcon from 'icons/full-archive.svg'
 import fullCopyIcon from 'icons/full-duplicate.svg'
 import fullPlusIcon from 'icons/full-plus.svg'
@@ -44,43 +40,27 @@ import fullShowIcon from 'icons/full-show.svg'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant } from 'ui-kit/Button/types'
-import { Divider } from 'ui-kit/Divider/Divider'
-import { Tabs } from 'ui-kit/Tabs/Tabs'
+import { Tab, Tabs } from 'ui-kit/Tabs/Tabs'
 
-import styles from './CollectiveOfferNavigation.module.scss'
+import { CollectiveOfferStep } from '../CollectiveOfferNavigation/CollectiveCreationOfferNavigation'
 
-export enum CollectiveOfferStep {
-  DETAILS = 'details',
-  STOCKS = 'stocks',
-  VISIBILITY = 'visibility',
-  SUMMARY = 'recapitulatif',
-  CONFIRMATION = 'confirmation',
-  PREVIEW = 'preview',
-}
+import styles from './CollectiveEditionOfferNavigation.module.scss'
 
-export interface CollectiveOfferNavigationProps {
+export interface CollectiveEditionOfferNavigationProps {
   activeStep: CollectiveOfferStep
-  isCreatingOffer: boolean
-  isCompletingDraft?: boolean
-  offerId?: number
-  className?: string
   isTemplate: boolean
-  requestId?: string | null
+  offerId?: number
   offer?:
     | GetCollectiveOfferResponseModel
     | GetCollectiveOfferTemplateResponseModel
 }
 
-export const CollectiveOfferNavigation = ({
+export const CollectiveEditionOfferNavigation = ({
   activeStep,
-  isCreatingOffer,
   isTemplate = false,
-  isCompletingDraft = false,
   offerId = 0,
-  className,
-  requestId = null,
   offer,
-}: CollectiveOfferNavigationProps): JSX.Element => {
+}: CollectiveEditionOfferNavigationProps): JSX.Element => {
   const { logEvent } = useAnalytics()
   const notify = useNotification()
   const navigate = useNavigate()
@@ -89,130 +69,12 @@ export const CollectiveOfferNavigation = ({
   const areCollectiveNewStatusesEnabled = useActiveFeature(
     'ENABLE_COLLECTIVE_NEW_STATUSES'
   )
-  const selectedOffererId = useSelector(selectCurrentOffererId)
 
-  const { mutate } = useSWRConfig()
+  const selectedOffererId = useSelector(selectCurrentOffererId)
 
   const [isArchiveModalOpen, setIsArchiveModalOpen] = useState(false)
 
   const id = computeURLCollectiveOfferId(offerId, Boolean(isTemplate))
-
-  const previewLink = `/offre/${id}/collectif${isTemplate ? '/vitrine' : ''}/apercu`
-
-  const stockEditionUrl = `/offre/${offerId}/collectif/stocks/edition`
-
-  const isEditingExistingOffer = !(isCreatingOffer || isCompletingDraft)
-
-  const stepList: { [key in CollectiveOfferStep]?: Step } = {}
-
-  const canEditOffer =
-    offer?.status !== CollectiveOfferStatus.ARCHIVED &&
-    location.pathname.includes('edition')
-
-  const canPreviewOffer =
-    (isCollectiveOffer(offer) &&
-      offer.displayedStatus !== CollectiveOfferDisplayedStatus.ARCHIVED) ||
-    !isCollectiveOffer(offer)
-
-  const requestIdUrl = requestId ? `?requete=${requestId}` : ''
-
-  if (isEditingExistingOffer) {
-    if (!isTemplate && canEditOffer) {
-      stepList[CollectiveOfferStep.DETAILS] = {
-        id: CollectiveOfferStep.DETAILS,
-        label: 'Détails de l’offre',
-        url: `/offre/${offerId}/collectif/edition`,
-      }
-      stepList[CollectiveOfferStep.STOCKS] = {
-        id: CollectiveOfferStep.STOCKS,
-        label: 'Dates et prix',
-        url: stockEditionUrl,
-      }
-      stepList[CollectiveOfferStep.VISIBILITY] = {
-        id: CollectiveOfferStep.VISIBILITY,
-        label: 'Établissement et enseignant',
-        url: `/offre/${offerId}/collectif/visibilite/edition`,
-      }
-    }
-  } else {
-    //  Creating an offer
-    stepList[CollectiveOfferStep.DETAILS] = {
-      id: CollectiveOfferStep.DETAILS,
-      label: 'Détails de l’offre',
-    }
-    if (!isTemplate) {
-      //  These steps only exist for bookable offers
-      stepList[CollectiveOfferStep.STOCKS] = {
-        id: CollectiveOfferStep.STOCKS,
-        label: 'Dates et prix',
-      }
-
-      stepList[CollectiveOfferStep.VISIBILITY] = {
-        id: CollectiveOfferStep.VISIBILITY,
-        label: 'Établissement et enseignant',
-      }
-    }
-    stepList[CollectiveOfferStep.SUMMARY] = {
-      id: CollectiveOfferStep.SUMMARY,
-      label: 'Récapitulatif',
-    }
-    stepList[CollectiveOfferStep.PREVIEW] = {
-      id: CollectiveOfferStep.PREVIEW,
-      label: 'Aperçu',
-    }
-    stepList[CollectiveOfferStep.CONFIRMATION] = {
-      id: CollectiveOfferStep.CONFIRMATION,
-      label: 'Confirmation',
-    }
-
-    const hasOfferPassedDetailsStep = offer && offerId
-    const hasOfferPassedStocksStep =
-      hasOfferPassedDetailsStep &&
-      (isCollectiveOfferTemplate(offer) || offer.collectiveStock)
-    const hasOfferPassedVisibilityStep =
-      hasOfferPassedStocksStep &&
-      (isCollectiveOfferTemplate(offer) || offer.institution)
-
-    if (hasOfferPassedDetailsStep) {
-      //  The user can go back to the details page after it has been filled the first time
-      stepList[CollectiveOfferStep.DETAILS].url = isTemplate
-        ? `/offre/collectif/vitrine/${offerId}/creation`
-        : `/offre/collectif/${offerId}/creation${requestIdUrl}`
-
-      if (
-        !isCollectiveOfferTemplate(offer) &&
-        stepList[CollectiveOfferStep.STOCKS]
-      ) {
-        //  The stocks step is accessible when the details form has been filled and the offer is bookable
-        stepList[CollectiveOfferStep.STOCKS].url =
-          `/offre/${offerId}/collectif/stocks`
-      }
-    }
-
-    if (hasOfferPassedStocksStep && stepList[CollectiveOfferStep.VISIBILITY]) {
-      //  The visibility tab is only accessible when the stocks form has been filled
-      stepList[CollectiveOfferStep.VISIBILITY].url =
-        `/offre/${offerId}/collectif/visibilite`
-    }
-
-    if (hasOfferPassedVisibilityStep) {
-      //  The summary tab is only accessible when the visibility form has been filled (or the offer is a template)
-      stepList[CollectiveOfferStep.SUMMARY].url = isTemplate
-        ? `/offre/${offerId}/collectif/vitrine/creation/recapitulatif`
-        : `/offre/${offerId}/collectif/creation/recapitulatif`
-
-      stepList[CollectiveOfferStep.PREVIEW].url = isTemplate
-        ? `/offre/${offerId}/collectif/vitrine/creation/apercu`
-        : `/offre/${offerId}/collectif/creation/apercu`
-    }
-  }
-
-  const steps = Object.values(stepList)
-  const tabs = steps.map(({ id, label, url }) => ({
-    key: id,
-    label,
-    url,
-  }))
 
   const archiveOffer = async () => {
     if (!offerId) {
@@ -222,16 +84,14 @@ export const CollectiveOfferNavigation = ({
     try {
       if (isTemplate) {
         await api.patchCollectiveOffersTemplateArchive({ ids: [offerId] })
-      } else {
-        await api.patchCollectiveOffersArchive({ ids: [offerId] })
-      }
-
-      if (isTemplate) {
         await mutate([GET_COLLECTIVE_OFFER_TEMPLATE_QUERY_KEY, offerId])
       } else {
+        await api.patchCollectiveOffersArchive({ ids: [offerId] })
         await mutate([GET_COLLECTIVE_OFFER_QUERY_KEY, offerId])
       }
+
       setIsArchiveModalOpen(false)
+
       notify.success('Une offre a bien été archivée', {
         duration: NOTIFICATION_LONG_SHOW_DURATION,
       })
@@ -241,6 +101,15 @@ export const CollectiveOfferNavigation = ({
       })
     }
   }
+
+  const canEditOffer =
+    offer?.status !== CollectiveOfferStatus.ARCHIVED &&
+    location.pathname.includes('edition')
+
+  const canPreviewOffer =
+    (isCollectiveOffer(offer) &&
+      offer.displayedStatus !== CollectiveOfferDisplayedStatus.ARCHIVED) ||
+    !isCollectiveOffer(offer)
 
   const canArchiveOffer = areCollectiveNewStatusesEnabled
     ? offer &&
@@ -271,11 +140,32 @@ export const CollectiveOfferNavigation = ({
         offer.displayedStatus !== CollectiveOfferDisplayedStatus.PENDING
     : false
 
-  return isEditingExistingOffer ? (
+  const tabs: Tab[] = [
+    {
+      key: CollectiveOfferStep.DETAILS,
+      label: 'Détails de l’offre',
+      url: `/offre/${offerId}/collectif/edition`,
+    },
+    {
+      key: CollectiveOfferStep.STOCKS,
+      label: 'Dates et prix',
+      url: `/offre/${offerId}/collectif/stocks/edition`,
+    },
+    {
+      key: CollectiveOfferStep.VISIBILITY,
+      label: 'Établissement et enseignant',
+      url: `/offre/${offerId}/collectif/visibilite/edition`,
+    },
+  ]
+
+  return (
     <>
       <div className={styles['duplicate-offer']}>
         {canPreviewOffer && (
-          <ButtonLink to={previewLink} icon={fullShowIcon}>
+          <ButtonLink
+            to={`/offre/${id}/collectif${isTemplate ? '/vitrine' : ''}/apercu`}
+            icon={fullShowIcon}
+          >
             Aperçu dans ADAGE
           </ButtonLink>
         )}
@@ -337,8 +227,7 @@ export const CollectiveOfferNavigation = ({
           </Button>
         )}
       </div>
-      <Divider />
-      {tabs.length > 0 && (
+      {!isTemplate && canEditOffer && (
         <Tabs
           tabs={tabs}
           selectedKey={activeStep}
@@ -358,7 +247,5 @@ export const CollectiveOfferNavigation = ({
         isDialogOpen={isArchiveModalOpen}
       />
     </>
-  ) : (
-    <Stepper activeStep={activeStep} className={className} steps={steps} />
   )
 }
