+++ pro/src/pages/CollectiveOffer/CollectiveOfferLayout/__specs__/CollectiveOfferLayout.spec.tsx
@@ -1,5 +1,4 @@
 import { screen } from '@testing-library/react'
-import React from 'react'
 
 vi.mock('apiClient/api', () => ({
   api: {
