+++ pro/src/pages/CollectiveOfferRoutes/utils/__specs__/getActiveStep.spec.tsx
@@ -1,4 +1,4 @@
-import { CollectiveOfferStep } from 'pages/CollectiveOffer/CollectiveOfferLayout/CollectiveOfferNavigation/CollectiveOfferNavigation'
+import { CollectiveOfferStep } from 'pages/CollectiveOffer/CollectiveOfferLayout/CollectiveOfferNavigation/CollectiveCreationOfferNavigation'
 
 import { getActiveStep } from '../getActiveStep'
